+++ frontend/src/App.tsx
@@ -2,7 +2,8 @@ import { Route, Routes } from "react-router-dom";
 
 import { AuthProvider } from "./auth/AuthContext";
 import { ProtectedRoute } from "./auth/ProtectedRoute";
-import Home from "./pages/Home";
+import Facilities from "./pages/Facilities";
+import FacilityAvailability from "./pages/FacilityAvailability";
 import SignIn from "./pages/SignIn";
 
 export default function App() {
@@ -10,14 +11,9 @@ export default function App() {
     <AuthProvider>
       <Routes>
         <Route path="/signin" element={<SignIn />} />
-        <Route
-          path="/"
-          element={
-            <ProtectedRoute>
-              <Home />
-            </ProtectedRoute>
-          }
-        />
+        <Route path="/" element={<ProtectedRoute><Facilities /></ProtectedRoute>} />
+        <Route path="/facilities/:facilityId"
+          element={<ProtectedRoute><FacilityAvailability /></ProtectedRoute>} />
       </Routes>
     </AuthProvider>
   );